# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * StyleSwipe API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface RegisterRequest {
  name: string;
  email: string;
}

export interface LoginRequest {
  email: string;
}

export interface AuthResponse {
  userId: number;
  name: string;
  email: string;
  token: string;
}

export interface ErrorResponse {
  error: string;
}

export interface StylePhoto {
  id: number;
  url: string;
  tags: string[];
  source?: string;
}

export interface StylePhotosResponse {
  photos: StylePhoto[];
  total: number;
}

export interface SwipeRequest {
  photoId: number;
  liked: boolean;
}

export interface SwipeResponse {
  success: boolean;
  swipeId: number;
}

export interface StyleTagWeight {
  tag: string;
  score: number;
  count: number;
}

export interface StyleProfileResponse {
  topTags: string[];
  tagWeights: StyleTagWeight[];
  totalSwipes: number;
  likedCount: number;
}

export interface StyleBoardResponse {
  photos: StylePhoto[];
}

export type GetStylePhotosParams = {
limit?: number;
offset?: number;
};