# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: StyleSwipe API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: auth
    description: Authentication operations
  - name: style
    description: Style quiz and profile operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /auth/register:
    post:
      operationId: registerUser
      tags: [auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterRequest"
      responses:
        "201":
          description: User created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "409":
          description: Email already registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /auth/login:
    post:
      operationId: loginUser
      tags: [auth]
      summary: Login with email
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "404":
          description: User not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /style-photos:
    get:
      operationId: getStylePhotos
      tags: [style]
      summary: Get style photos for the quiz
      security:
        - bearerAuth: []
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        "200":
          description: List of style photos not yet swiped by the user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StylePhotosResponse"

  /swipes:
    post:
      operationId: recordSwipe
      tags: [style]
      summary: Record a swipe on a style photo
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SwipeRequest"
      responses:
        "201":
          description: Swipe recorded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SwipeResponse"

  /style-profile:
    get:
      operationId: getStyleProfile
      tags: [style]
      summary: Get the user's computed style profile
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Style profile with weighted tags
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StyleProfileResponse"

  /style-board:
    get:
      operationId: getStyleBoard
      tags: [style]
      summary: Get the user's liked photos (style board)
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Liked photos
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StyleBoardResponse"

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer

  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    RegisterRequest:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
      required:
        - name
        - email

    LoginRequest:
      type: object
      properties:
        email:
          type: string
      required:
        - email

    AuthResponse:
      type: object
      properties:
        userId:
          type: integer
        name:
          type: string
        email:
          type: string
        token:
          type: string
      required:
        - userId
        - name
        - email
        - token

    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error

    StylePhoto:
      type: object
      properties:
        id:
          type: integer
        url:
          type: string
        tags:
          type: array
          items:
            type: string
        source:
          type: string
      required:
        - id
        - url
        - tags

    StylePhotosResponse:
      type: object
      properties:
        photos:
          type: array
          items:
            $ref: "#/components/schemas/StylePhoto"
        total:
          type: integer
      required:
        - photos
        - total

    SwipeRequest:
      type: object
      properties:
        photoId:
          type: integer
        liked:
          type: boolean
      required:
        - photoId
        - liked

    SwipeResponse:
      type: object
      properties:
        success:
          type: boolean
        swipeId:
          type: integer
      required:
        - success
        - swipeId

    StyleTagWeight:
      type: object
      properties:
        tag:
          type: string
        score:
          type: number
        count:
          type: integer
      required:
        - tag
        - score
        - count

    StyleProfileResponse:
      type: object
      properties:
        topTags:
          type: array
          items:
            type: string
        tagWeights:
          type: array
          items:
            $ref: "#/components/schemas/StyleTagWeight"
        totalSwipes:
          type: integer
        likedCount:
          type: integer
      required:
        - topTags
        - tagWeights
        - totalSwipes
        - likedCount

    StyleBoardResponse:
      type: object
      properties:
        photos:
          type: array
          items:
            $ref: "#/components/schemas/StylePhoto"
      required:
        - photos