# lib/api-zod/src/generated/types/stylePhotosResponse.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * StyleSwipe API specification
 * OpenAPI spec version: 0.1.0
 */
import type { StylePhoto } from './stylePhoto';

export interface StylePhotosResponse {
  photos: StylePhoto[];
  total: number;
}