# lib/api-zod/src/generated/types/styleProfileResponse.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * StyleSwipe API specification
 * OpenAPI spec version: 0.1.0
 */
import type { StyleTagWeight } from './styleTagWeight';

export interface StyleProfileResponse {
  topTags: string[];
  tagWeights: StyleTagWeight[];
  totalSwipes: number;
  likedCount: number;
}