# lib/api-zod/src/generated/types/swipeRequest.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * StyleSwipe API specification
 * OpenAPI spec version: 0.1.0
 */

export interface SwipeRequest {
  photoId: number;
  liked: boolean;
}